+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Add transition callback. A function which is called for every transition. It has default empty
 implementation.
+- Add support for implicit and wildcard internal transitions
+
+### Changed
+
+- [breaking] Renamed custom_guard_error flag to custom_error as it is not guard specific anymore
+- [breaking] Re-ordered on_exit/on_entry hooks calls
 
 ## [0.7.0] - 2024-07-03